+++ pkg/resourcemanager/mock/eventhubs/managers.go
