+++ pkg/resourcemanager/mock/storages/storage.go
@@ -21,7 +21,7 @@ import (
 	"errors"
 	"github.com/Azure/azure-sdk-for-go/services/storage/mgmt/2019-04-01/storage"
 	apiv1 "github.com/Azure/azure-service-operator/api/v1"
-	"github.com/Azure/azure-service-operator/pkg/resourcemanager/mock_test/helpers"
+	"github.com/Azure/azure-service-operator/pkg/resourcemanager/mock/helpers"
 	"github.com/Azure/go-autorest/autorest"
 	"github.com/Azure/go-autorest/autorest/to"
 )
